+++ RELEASE_NOTES.md
@@ -2,8 +2,10 @@ Release Notes
 =============
 
 ## 1.6.13
+* Alerts: Initial support for Alerts
 * Container Groups: Fix to generate parameters for secure environment variables on `initContainers`.
 * Key Vaults: Support for adding access policies on an existing key vault with `keyVaultAddPolicies`.
+* Virtual Networks: support for adding subnets to existing virtual networks.
 
 ## 1.6.12
 * Custom FarmerException raised for all exceptions.